// spring-ai-alibaba-studio/spring-ai-alibaba-studio-server/frontend/packages/spark-flow/src/demos/flow-demo/types/flow.ts
export interface IModelConfigParamItem {
  key: string;
  type: IValueType;
  value: number | string | boolean;
  enable: boolean;
}

export interface IModelVisionConfig {
  enable: boolean;
  params: INodeDataInputParamItem[];
}

export interface ISelectedModelParams {
  model_id: string;
  model_name: string;
  provider: string;
  params: IModelConfigParamItem[];
  mode: 'chat' | 'completion';
  vision_config: IModelVisionConfig;
}

export interface IShortMemoryConfig {
  enabled: boolean;
  /* Memory type, self: cache for this node, custom: custom cache */
  type: 'self' | 'custom';
  /* Memory window, when type is self, memory window needs to be defined */
  round?: number;
  param: INodeDataInputParamItem;
}

export interface IRetryConfig {
  /* Whether to enable retry */
  retry_enabled: boolean;
  /* Maximum number of retries */
  max_retries: number;
  /* Retry interval */
  retry_interval: number;
}

export interface ITryCatchConfig {
  /* Exception handling strategy, defaultValue: default value, failBranch: failure branch, noop: do nothing */
  strategy: 'defaultValue' | 'failBranch' | 'noop';
  /* If strategy is defaultValue, default_value needs to be defined */
  default_values?: Omit<INodeDataInputParamItem, 'value_from'>[];
}

export interface ILLMNodeParam {
  /* System prompt word */
  sys_prompt_content?: string;
  prompt_content?: string;
  model_config: ISelectedModelParams;
  /* Short-term memory */
  short_memory: IShortMemoryConfig;
  /* Retry configuration */
  retry_config: IRetryConfig;
  /* Exception handling configuration */
  try_catch_config: ITryCatchConfig;
}
// ...
export type INodeDataNodeParam =
  | IStartNodeParam
  | IInputNodeParam
  | IEndNodeParam
  | IOutputNodeParam
  | ILLMNodeParam
  | IParameterExtractorNodeParam
  | IJudgeNodeParam
  | IApiNodeParam
  | IParallelNodeParam
  | IIteratorNodeParam
  | IAppComponentNodeParam
  | IPluginNodeParam
  | IClassifierNodeParam
  | IScriptNodeParam
  | IVariableAssignNodeParam
  | IVariableHandleNodeParam
  | IRetrievalNodeParam
  | IMCPNodeParam;

export type IWorkFlowNodeData<T = INodeDataNodeParam> = {
  /* Node name */
  label: string;
  /* Node input parameters */
  input_params: INodeDataInputParamItem[];
  /* Node output parameters */
  output_params: INodeDataOutputParamItem[];
  /* Node business parameters */
  node_param: T;
  /* Node description */
  desc?: string;
};

export type IStartNodeData = IWorkFlowNodeData<IStartNodeParam>;

export type IInputNodeData = IWorkFlowNodeData<IInputNodeParam>;

export type IEndNodeData = IWorkFlowNodeData<IEndNodeParam>;

export type IOutputNodeData = IWorkFlowNodeData<IOutputNodeParam>;

export type ILLMNodeData = IWorkFlowNodeData<ILLMNodeParam>;